Refactor: priority_queue -> rabbit_mq_utils, improve README.md
