+++ docker-compose.yml
@@ -1,7 +1,7 @@
 version: '3.8'
 
 services:
-  priority-queue:
+  operandi-rabbit-mq:
     image: rabbit-mq
     container_name: operandi-rabbit-mq
     hostname: rabbit-mq-host
@@ -11,6 +11,7 @@ services:
     ports:
       - "5672:5672"
       - "15672:15672"
+      - "25672:25672"
     volumes:
       - ~/.docker-conf/rabbitmq/etc/:/var/etc/rabbitmq/
       - ~/.docker-conf/rabbitmq/data/:/var/lib/rabbitmq/
@@ -38,7 +39,7 @@ services:
       context: ./src
       dockerfile: ./Dockerfile_server  
     depends_on:
-      priority-queue:
+      operandi-rabbit-mq:
         condition: service_healthy
     restart: on-failure
     ports:
@@ -54,7 +55,7 @@ services:
       context: ./src
       dockerfile: ./Dockerfile_broker
     depends_on:
-      priority-queue:
+      operandi-rabbit-mq:
         condition: service_healthy
     restart: on-failure
     networks:
