+++ src/Dockerfile_broker
@@ -7,7 +7,7 @@ ENV LC_ALL=C.UTF-8
 ENV LANG=C.UTF-8
 
 COPY ./service_broker/ /usr/src/service_broker
-COPY ./priority_queue/ /usr/src/priority_queue
+COPY ./rabbit_mq_utils/ /usr/src/rabbit_mq_utils
 
 WORKDIR /usr/src/service_broker/service_broker
 
@@ -15,6 +15,7 @@ WORKDIR /usr/src/service_broker/service_broker
 RUN apt-get update && apt-get -y install \
     ca-certificates \
     curl \
+    git  \
     iputils-ping \
     make \
     net-tools \
@@ -31,8 +32,7 @@ RUN python3 -m pip install --upgrade pip setuptools
 RUN pip3 install -U pip wheel
 RUN python3 -m pip install -r /usr/src/service_broker/requirements.txt
 RUN pip3 install /usr/src/service_broker
-RUN python3 -m pip install -r /usr/src/service_broker/requirements.txt
-RUN pip3 install /usr/src/priority_queue
+RUN python3 -m pip install -r /usr/src/rabbit_mq_utils/requirements.txt
+RUN pip3 install /usr/src/rabbit_mq_utils
 
 RUN operandi-broker --version
-
