+++ src/Dockerfile_rabbitmq
@@ -1,13 +1,10 @@
 FROM rabbitmq:3.8.27-management-alpine
 
-ADD ./priority_queue/rabbitmq.conf /etc/rabbitmq/
-# ADD ./priority_queue/definitions.json /etc/rabbitmq/
+ADD ./rabbit_mq_utils/rabbitmq.conf /etc/rabbitmq/
+# ADD ./rabbit_mq_utils/definitions.json /etc/rabbitmq/
 
 RUN chown rabbitmq:rabbitmq /etc/rabbitmq/rabbitmq.conf /etc/rabbitmq/rabbitmq.conf
 # RUN chown rabbitmq:rabbitmq /etc/rabbitmq/definitions.json /etc/rabbitmq/definitions.json
 
-ADD --chown=rabbitmq ./priority_queue/definitions.json /etc/rabbitmq/
+ADD --chown=rabbitmq ./rabbit_mq_utils/definitions.json /etc/rabbitmq/
 ENV RABBITMQ_SERVER_ADDITIONAL_ERL_ARGS="-rabbitmq_management load_definitions \"/etc/rabbitmq/definitions.json\""
-
-EXPOSE 5672 15672
-
