+++ src/Dockerfile_server
@@ -7,7 +7,7 @@ ENV LC_ALL=C.UTF-8
 ENV LANG=C.UTF-8
 
 COPY ./operandi_server/ /usr/src/operandi_server
-COPY ./priority_queue/ /usr/src/priority_queue
+COPY ./rabbit_mq_utils/ /usr/src/rabbit_mq_utils
 
 WORKDIR /usr/src/operandi_server/operandi_server
 
@@ -15,6 +15,7 @@ WORKDIR /usr/src/operandi_server/operandi_server
 RUN apt-get update && apt-get -y install \
     ca-certificates \
     curl \
+    git  \
     iputils-ping \
     make \
     net-tools \
@@ -31,10 +32,9 @@ RUN python3 -m pip install --upgrade pip setuptools
 RUN pip3 install -U pip wheel
 RUN python3 -m pip install -r /usr/src/operandi_server/requirements.txt
 RUN pip3 install /usr/src/operandi_server
-RUN python3 -m pip install -r /usr/src/priority_queue/requirements.txt
-RUN pip3 install /usr/src/priority_queue
+RUN python3 -m pip install -r /usr/src/rabbit_mq_utils/requirements.txt
+RUN pip3 install /usr/src/rabbit_mq_utils
 
 EXPOSE 8000
 
 RUN operandi-server --version
-
