+++ src/operandi_server/requirements.txt
@@ -6,4 +6,3 @@ pytest>=7.0.0
 click >=7
 tomli>=2.0.0
 ocrd_webapi @ git+https://github.com/OCR-D/ocrd-webapi-implementation.git@refactoring#egg=ocrd_webapi
-
