+++ src/operandi_server/setup.py
@@ -6,7 +6,7 @@ install_requires = open('requirements.txt').read().split('\n')
 setup(
     name='operandi_server',
     version='1.1.1',
-    description='OPERANDI - REST API Server',
+    description='OPERANDI - Server',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
     author='Mehmed Mustafa',
