+++ src/rabbit_mq_utils/definitions.json
