+++ src/rabbit_mq_utils/rabbit_mq_utils/constants.py
