+++ src/rabbit_mq_utils/requirements.txt
