+++ src/service_broker/requirements.txt
@@ -5,4 +5,3 @@ pydantic>=1.9.1
 pytest>=7.0.0
 click>=7
 tomli>=2.0.0
-
